// src/app/organization/models/BasicDetails.ts
export class BasicDetails {
    orgHash = 'ORGANIZATION';
    org_name: string;
    orgKey?: string;
    org_type: string;
    institute_type: string
    org_logo: string;
    org_parent_id: string;
    seller_id: string;
    org_short_code: string;
    org_code: string;
    isDeleted?: boolean;
    isActivated?: boolean;
}